pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "py_document_chunker"
version = "0.1.0"
description = "A state-of-the-art Python package for advanced text segmentation (chunking)."
readme = { file = "README.md", content-type = "text/markdown" }
keywords = ["text chunking", "text splitting", "RAG", "Retrieval-Augmented Generation", "NLP", "natural language processing", "document processing", "langchain", "llamaindex"]
authors = [{ name = "Gowtham Rao", email = "rao@ohdsi.org" }]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Topic :: Text Processing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
requires-python = ">=3.8"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "lxml>=5.2.0"
]

[project.optional-dependencies]
# R-6.3.2: Optional dependencies for specialized features
nlp = [
    "nltk>3.8.1",
]
spacy = [
    "spacy>=3.7.0",
]
semantic = [
    "numpy>=2.0.0",
]
code = [
    "tree-sitter>=0.21.0",
    "tree-sitter-language-pack>=0.3.0"
]
markdown = [
    "markdown-it-py>=3.0.0"
]
html = [
    "html5lib>=1.1",
]
# For integration with other frameworks
langchain = ["langchain-core"]
llamaindex = ["llama-index-core>=0.12.29"]
tokenizers = [
    "tiktoken>=0.7.0"
]
# For development and testing
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "py_document_chunker[tokenizers,nlp,spacy,semantic,code,markdown,langchain,llamaindex,html]",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0",
]


[project.urls]
"Homepage" = "https://github.com/OHDSI/py_document_chunker"
"Bug Tracker" = "https://github.com/OHDSI/py_document_chunker/issues"

[tool.pytest.ini_options]
pythonpath = "src"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
